requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
flake8==7.1.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.2
pluggy==1.4.0
pycodestyle==2.12.0
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
pytest==8.0.0
requests==2.31.0
typing_extensions==4.12.2
urllib3==2.2.0
